[FrameworkBundle] Extends the RequestDataCollector
Moves the logic of collect controller data while forwarding
in the framework bundle.
